points per dollar added to flex and superflex
