--- model/dfs_positions_calc_funcs/stackingValuesCalcs/bestTripleStackWithQBWROne.js
@@ -1,10 +1,0 @@
-'use strict';
-
-const allQBData = require('../qbValuesCalcs');
-const allRBData = require('../rbValuesCalcs');
-const allWRData = require('../wrValuesCalcs');
-const allTEData = require('../teValuesCalcs');
-const allQBs = require('../../teamandpositionvariables/allQBVariables');
-const allRBs = require('../../teamandpositionvariables/allRBVariables');
-const allWRs = require('../../teamandpositionvariables/allWRVariables');
-const allTEs = require('../../teamandpositionvariables/allTEVariables');
